pyproject.toml
[project]
name = "paper-radar"
version = "0.1.0"
description = ""
authors = [
    {name = "MorganREN",email = "mohan.ren@uq.edu.au"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0.0"
dependencies = [
    "sqlmodel (>=0.0.31,<0.0.32)",
    "arxiv (>=2.4.0,<3.0.0)",
    "openai (>=2.15.0,<3.0.0)",
    "langgraph (>=1.0.7,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "playwright (>=1.57.0,<2.0.0)",
    "playwright-stealth (>=2.0.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "fake-useragent (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
